+++ djangocms_frontend/contrib/grid/frameworks/bootstrap5.py
@@ -79,7 +79,7 @@ class GridContainerFormMixin(EntangledModelFormMixin):
         choices=settings.EMPTY_CHOICE + settings.framework_settings.SHADOW_CHOICES,
         initial=settings.EMPTY_CHOICE[0][0],
         widget=ButtonGroup(attrs=dict(property="shadow")),
-        help_text=_("Use shadows to optically lift cards up from the background."),
+        help_text=_("Use shadows to optically lift cards from the background."),
     )