+++ djangocms_frontend/contrib/image/models.py
@@ -2,9 +2,9 @@ from django.conf import settings
 from django.utils.functional import cached_property
 from django.utils.translation import gettext as _
 from easy_thumbnails.files import get_thumbnailer
-from entangled.utils import get_related_object
 
 from djangocms_frontend.contrib.link.models import GetLinkMixin
+from djangocms_frontend.helpers import get_related_object
 from djangocms_frontend.models import FrontendUIItem
 
 # use golden ration as default (https://en.wikipedia.org/wiki/Golden_ratio)